Rename range-node-conversion.ts → to-range.ts
