+++ packages/dom/src/text-quote/describe.ts
@@ -24,7 +24,7 @@ import type {
 } from '@apache-annotator/selector';
 import { describeTextQuote as abstractDescribeTextQuote } from '@apache-annotator/selector';
 import { ownerDocument } from '../owner-document';
-import { toRange } from '../range-node-conversion';
+import { toRange } from '../to-range';
 import { TextNodeChunker } from '../text-node-chunker';
 
 /**
