+++ packages/dom/src/to-range.ts
